Add files via upload
In this project, we forecasted the weekly prices of conventional avocado’s of Westren United States through Auto ARIMA and Autoregressive Integretated Moving Average (ARIMA) models. The data has several missing values which are delt with linear in- terpolation. The stationarity of data is tested through Augmented Dickey–Fuller and KPSS tests which shows that the data is non stationary. The visulaization shows that the prices are not normal and the data set has some outliers. The data is split into training and testing data in which the testing data part includes first the ten weeks of 2018. Several ARIMA models are built and their order is identified by the ACF and PACF plots. The ARIMA model that provided the lowest AIC value was therefore used to forecast the avocado prices. The performance of these models was tested using ac- curacy metrics such as ME, MAE, RMSE, MSE, and MAPE. The results indicate that the ARIMA(0,1,1)(0,0,1) model performs better than the Auto ARIMA(1,1,0)(1,0,0) model.